requirements.txt
# Core LangChain (pinned to avoid breaking changes)
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
pypdf>=4.0.0
python-dotenv
openai

# Backend
fastapi
uvicorn[standard]
python-multipart

# Frontend
streamlit

# Utils
requests
httpx
tqdm
pydantic
pytesseract 
pdf2image